% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict18.R
\docType{data}
\name{dict18}
\alias{dict18}
\title{Diccionario para los datos de las encuestas ENHOGAR}
\format{
\link{list}
}
\source{
\url{https://github.com/endomer/enhogar/tree/master/Files/Cuestionarios/}
}
\usage{
dict18
}
\description{
Contiene las etiquetas tanto para las variables como para los valores de estas.
}
\keyword{datasets}
