% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_browse_dict.R
\name{ehg_browse_dict}
\alias{ehg_browse_dict}
\title{Visualizar el diccionario de forma interactiva
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}`}
\usage{
ehg_browse_dict(edition = get_enhogar_edition())
}
\arguments{
\item{edition}{\link{integer}: Edición de la encuesta. Vea \code{get_enhogar_edition()}
para más información.}
}
\value{
Widget interactivo con el diccionario de etiquetas
}
\description{
Visualizar el diccionario de forma interactiva
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}`
}
\examples{
\dontrun{
  ehg_browse_dict()
}
}
