% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_dbConnect.R
\name{ehg_dbConnect}
\alias{ehg_dbConnect}
\title{Conexión a base de datos}
\usage{
ehg_dbConnect()
}
\value{
Conexión a base de datos.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Vea \code{Dmisc::\link[Dmisc:dbConnect]{dbConnect}}
}
\examples{
\dontrun{
  conn <- ehg_dbConnect()
}
}
