% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_desocupado.R
\name{ehg_desocupado}
\alias{ehg_desocupado}
\title{Población desocupada}
\usage{
ehg_desocupado(tbl, min_edad = 15, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: dataset con los datos de la encuesta}

\item{min_edad}{\link{numeric}: Edad mínima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}

\item{max_edad}{\link{numeric}: Edad máxima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}
}
\value{
\link{data.frame}: los datos suministrados en el input \code{tbl} con las
variables \code{pet}, \code{ocupado} y \code{desocupado} agregadas.
}
\description{
Desocupación Abierta (PDAb): Población de 10 años y más que en el período
de referencia declaró no tener trabajo, que están disponibles para trabajar
de inmediato y que en las últimas cuatro semanas han realizado diligencias
para buscar un trabajo.
\strong{Glosario de términos - Banco Central de la República Dominicana}
}
\details{
Los datos de la ENHOGAR no permiten aislar las personas que aunque buscaron
trabajo no estaban disponibles para comenzar a trabajar, que según la
definición no forman parte de la desocupación abierta, y forman parte de la
Fuerza de trabajo potencial (Vea \code{\link{ehg_fuerza_trabajo_potencial}}).
Dado que los primeros son más, se incluyen todos los que buscaron trabajo en
la población desocupada, reduciendo con ello la fuerza de trabajo potencial.
}
\examples{
\dontrun{
  enhogar <- ehg_desocupado(enhogar)
}
}
