% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_fuerza_trabajo_potencial.R
\name{ehg_fuerza_trabajo_potencial}
\alias{ehg_fuerza_trabajo_potencial}
\title{Fuerza de trabajo potencial
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ehg_fuerza_trabajo_potencial(tbl, min_edad = 15, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: dataset con los datos de la encuesta}

\item{min_edad}{\link{numeric}: Edad mínima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}

\item{max_edad}{\link{numeric}: Edad máxima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}
}
\value{
\link{data.frame}: los datos suministrados en el input \code{tbl} con las
variables \code{pet}, \code{ocupado}, \code{desocupado}, \code{pea} y \code{fuerza_trabajo_potencial} agregadas.
}
\description{
Abarque a las personas que hayan manifestado su interés en trabajar,
y que actualmente se contabilizan como personas fuera de la fuerza de trabajo.
Distingue tres grupos mutuamente excluyentes:
\itemize{
\item{
a) buscadores no disponibles: se refiere a personas sin empleo que están
en busca de empleo pero no se encuentran disponibles;
}
\item{
b) buscadores potenciales disponibles: se refiere a las personas sin empleo
que no están en busca de empleo pero se encuentran disponibles para trabajar; y
}
\item{
c) buscadores potenciales que desean trabajar: se refiere a las personas
sin empleo que no están en busca de empleo ni se encuentran disponibles pero
que desean trabajar.
}
}
\strong{Key Indicators of the Labour Market (KILM) - Organización Internacional del
Trabajo (OIT), 2015}
}
\details{
Importante!
Vea la sección \code{detalles} de \code{\link{ehg_desocupado}} para más detalles.
}
\examples{
\dontrun{
  enhogar <- ehg_fuerza_trabajo_potencial(enhogar)
}
}
