% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_inactivo.R
\name{ehg_inactivo}
\alias{ehg_inactivo}
\title{Población No Económicamente Activa (PNEA) - Inactivos
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ehg_inactivo(tbl, min_edad = 15, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: dataset con los datos de la encuesta}

\item{min_edad}{\link{numeric}: Edad mínima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}

\item{max_edad}{\link{numeric}: Edad máxima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}
}
\value{
\link{data.frame}: los datos suministrados en el input \code{tbl} con las
variables \code{pet}, \code{ocupado}, \code{desocupado}, \code{pea} e \code{inactivo} agregadas.
}
\description{
Población No Económicamente Activa: Porción de la población en edad de
trabajar que no trabaja, no buscó trabajo activamente en las últimas cuatro
semanas, ni esta disponible para el trabajo; se incluye a los estudiantes,
las amas de casa, los rentistas, jubilados, los incapacitados para trabajar
y los vagos.
\strong{Glosario de términos - Banco Central de la República Dominicana}
}
\examples{
\dontrun{
  enhogar <- ehg_fuerza_trabajo_potencial(enhogar)
}
}
