% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_ocupado.R
\name{ehg_ocupado}
\alias{ehg_ocupado}
\title{Población ocupada}
\usage{
ehg_ocupado(tbl, min_edad = 15, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: dataset con los datos de la encuesta}

\item{min_edad}{\link{numeric}: Edad mínima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}

\item{max_edad}{\link{numeric}: Edad máxima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}
}
\value{
\link{data.frame}: los datos suministrados en el input \code{tbl} con las
variables \code{pet} y \code{ocupado} agregadas.
}
\description{
Población Ocupada (PO): Personas en edad de trabajar \code{\link{ehg_pet}},
que labora por lo menos una hora en el período de referencia.
Esta categoría incluye a todas aquellas personas de 10 años y más que
realizan una actividad económica y además, a las que al momento de la
encuesta, teniendo una ocupación, no asisten a su lugar de trabajo por
razones circunstanciales tales como: vacaciones, enfermedad, huelga, etc.
\strong{Glosario de términos - Banco Central de la República Dominicana}
}
\examples{
\dontrun{
  enhogar <- ehg_ocupado(enhogar)
}
}
