% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_pea.R
\name{ehg_pea}
\alias{ehg_pea}
\title{Población Económicamente Activa (PEA)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ehg_pea(tbl, min_edad = 15, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: dataset con los datos de la encuesta}

\item{min_edad}{\link{numeric}: Edad mínima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}

\item{max_edad}{\link{numeric}: Edad máxima a la que está permitido trabajar. Vea
\code{\link{ehg_pet}} para más detalles.}
}
\value{
\link{data.frame}: los datos suministrados en el input \code{tbl} con las
variables \code{pet}, \code{ocupado}, \code{desocupado} y \code{pea} agregadas.
}
\description{
Población No Económicamente Activa: Porción de la población en edad de
trabajar que no trabaja, no buscó trabajo activamente en las últimas
cuatro semanas, ni esta disponible para el trabajo; se incluye a los
estudiantes, las amas de casa, los rentistas, jubilados, los incapacitados
para trabajar y los vagos.
\strong{Glosario de términos - Banco Central de la República Dominicana}
}
\examples{
\dontrun{
  enhogar <- ehg_pea(enhogar)
}
}
