% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_pet.R
\name{ehg_pet}
\alias{ehg_pet}
\title{Población en Edad de Trabajar (PET)}
\usage{
ehg_pet(tbl, min_edad = 15, max_edad = Inf)
}
\arguments{
\item{tbl}{\link{data.frame}: el dataset con los datos}

\item{min_edad}{\link{numeric}: edad a partir de la cual se considera en edad de trabajar}

\item{max_edad}{\link{numeric}: edad hasta la que se considera en edad de trabajar}
}
\value{
\link{data.frame}: los datos suministrados en el input \code{tbl} con la
variable \code{pet} agregada.
}
\description{
Comprende todas las personas en edad legalmente aceptada para trabajar.
En el caso de la República Dominicana la edad mínima es 15 años, y no existe
una edad máxima.
}
\examples{
\dontrun{
  enhogar <- ehg_pet(enhogar)
}
}
