% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_Labels.R
\name{ehg_set_labels}
\alias{ehg_set_labels}
\alias{ehg_setLabels}
\title{Asigna etiquetas de datos a las variables especificadas
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ehg_set_labels(tbl, vars = NULL)

ehg_setLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Los datos introducidos en el argumento \code{tbl} pero con etiquetas de datos
}
\description{
Asigna etiquetas de datos a las variables especificadas
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
Esta función permite asignar etiquetas de datos a las variables de un data.frame
o equivalente, a partir del diccionario de la encuesta.
}
\examples{
\dontrun{
  ehg18 <- data.frame(ZONA = c(1, 2))
  str(ehg18)
  str(ehg_set_labels(ehg18))
}
}
\seealso{
Etiquetas de datos \code{vignette("labeler", package = "labeler")}
}
