% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_Labels.R
\name{ehg_setLabels}
\alias{ehg_setLabels}
\title{Asigna etiquetas de datos a las variables especificadas}
\usage{
ehg_setLabels(tbl, vars = NULL, year = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{year}{\link{numeric}: año de la encuesta con la que está trabajando.
\code{null} por defecto, y la función intenta determinar el año a partir de la
variable \code{"HANO"} si está presente en los datos}
}
\value{
Los datos introducidos en el argumento \code{tbl} pero con etiquetas de datos
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\details{
Esta función permite asignar etiquetas de datos a las variables de un data.frame
o equivalente, a partir del diccionario de la encuesta.
}
\examples{
\dontrun{
  ehg18 <- data.frame(ZONA = c(1, 2))
  str(ehg18)
  str(ftc_setLabels(ehg18))
}
}
\seealso{
Etiquetas de datos \code{vignette("labels", package = "endomer")}
}
