% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_Labels.R
\name{ehg_use_labels}
\alias{ehg_use_labels}
\alias{ehg_useLabels}
\title{Utiliza las etiquetas de datos de una variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
ehg_use_labels(tbl, vars = NULL)

ehg_useLabels(tbl, vars = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}
}
\value{
Los datos suministrados en el argumento \code{tbl}, pero en lugar de
valores utilizando las etiquetas de datos correspondientes
}
\description{
Utiliza las etiquetas de datos de una variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  ehg18 <- data.frame(ZONA = c(1, 2))
  ehg18
  ehg_use_labels(ehg18)
}
}
\seealso{
Etiquetas de datos \code{vignette("labeler", package = "labeler")}
}
