% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehg_Labels.R
\name{ehg_useLabels}
\alias{ehg_useLabels}
\title{Utiliza las etiquetas de datos de una variable}
\usage{
ehg_useLabels(tbl, vars = NULL, year = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexión a base de datos o dataframe con los datos}

\item{vars}{\link{character}: Si especificado, solo se asignaran las etiquetas a esas variables.}

\item{year}{\link{numeric}: año de la encuesta con la que está trabajando.
\code{null} por defecto, y la función intenta determinar el año a partir de la
variable \code{"HANO"} si está presente en los datos}
}
\value{
Los datos suministrados en el argumento \code{tbl}, pero en lugar de
valores utilizando las etiquetas de datos correspondientes
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
}
\examples{
\dontrun{
  ehg18 <- data.frame(ZONA = c(1, 2))
  ehg18
  ftc_useLabels(ehg18)
}
}
\seealso{
Etiquetas de datos \code{vignette("labels", package = "endomer")}
}
