% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhogar_edition.R
\name{enhogar_edition}
\alias{enhogar_edition}
\title{Edición de la encuesta
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
enhogar_edition(year)
}
\arguments{
\item{year}{\link{integer}: Año de la encuesta. Es equivalente a la edición.
Corresponde a un valor entre \code{2005} y \code{2019}. Años en los que
se ha realizado la encuesta.}
}
\value{
\link{NULL} Configura la opción ENHOGAR_EDITION. No devuelve nada.
}
\description{
Edición de la encuesta
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
 enhogar_edition(2019)
}
}
