% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhogar_edition.R
\name{get_enhogar_edition}
\alias{get_enhogar_edition}
\title{Obtiene la edici贸n de la encuesta actualmente configurada
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
get_enhogar_edition(tbl = NULL)
}
\arguments{
\item{tbl}{\link{data.frame}: NULL por defecto.
Conexi贸n a base de datos o dataframe con los datos si se suministra.}
}
\value{
\link{integer}: Edici贸n de la encuesta.
}
\description{
Obtiene la edici贸n de la encuesta actualmente configurada
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
 get_enhogar_edition()
}
}
