% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhogar_edition.R
\name{guess_enhogar_edition}
\alias{guess_enhogar_edition}
\title{Infiere la edici贸n de la encuesta a partir de los datos suministrados
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
guess_enhogar_edition(tbl)
}
\arguments{
\item{tbl}{\link{data.frame}: Conexi贸n a base de datos o dataframe con los datos}
}
\value{
\link{integer}: Edici贸n de la encuesta.
}
\description{
Infiere la edici贸n de la encuesta a partir de los datos suministrados
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
guess_enhogar_edition(enhogar_data)
}
}
